RUMM-2996: Move main WebView RUM and Logs features files into the dedicated module, code is not compileable
