--- library/dd-sdk-android-webview/src/main/kotlin/com/datadog/android/webview/Placeholder.kt
@@ -1,12 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2016-Present Datadog, Inc.
- */
-
-package com.datadog.android.webview
-
-/**
- * Empty temporary class to make Gradle tasks happy.
- */
-class Placeholder
