+++ sample/kotlin/build.gradle.kts
@@ -115,6 +115,7 @@ dependencies {
     implementation(project(":library:dd-sdk-android-session-replay"))
     implementation(project(":library:dd-sdk-android-ndk"))
     implementation(project(":library:dd-sdk-android-trace"))
+    implementation(project(":library:dd-sdk-android-webview"))
     implementation(project(":dd-sdk-android-ktx"))
     implementation(project(":dd-sdk-android-rx"))
     implementation(project(":dd-sdk-android-timber"))
